Update entry point to be lib/function_name.rb (more like a gaem)
- demo-function => demo_function
