+++ source/demo_function/README.md
@@ -1,4 +1,4 @@
-# demo-function
+# demo_function
 
 This project contains source code and supporting files for a serverless application that you can deploy with the SAM CLI. It includes the following files and folders.
 
@@ -49,7 +49,7 @@ You can find your API Gateway Endpoint URL in the output values displayed after 
 Build your application with the `sam build` command.
 
 ```bash
-demo-function$ sam build
+demo_function$ sam build
 ```
 
 The SAM CLI installs dependencies defined in `src/Gemfile`, creates a deployment package, and saves it in the `.aws-sam/build` folder.
@@ -59,14 +59,14 @@ Test a single function by invoking it directly with a test event. An event is a 
 Run functions locally and invoke them with the `sam local invoke` command.
 
 ```bash
-demo-function$ sam local invoke DemoFunctionFunction --event events/event.json
+demo_function$ sam local invoke DemoFunctionFunction --event events/event.json
 ```
 
 The SAM CLI can also emulate your application's API. Use the `sam local start-api` to run the API locally on port 3000.
 
 ```bash
-demo-function$ sam local start-api
-demo-function$ curl http://localhost:3000/
+demo_function$ sam local start-api
+demo_function$ curl http://localhost:3000/
 ```
 
 The SAM CLI reads the application template to determine the API's routes and the functions that they invoke. The `Events` property on each function's definition includes the route and method for each path.
@@ -90,7 +90,7 @@ To simplify troubleshooting, SAM CLI has a command called `sam logs`. `sam logs`
 `NOTE`: This command works for all AWS Lambda functions; not just the ones you deploy using SAM.
 
 ```bash
-demo-function$ sam logs -n DemoFunctionFunction --stack-name demo-function --tail
+demo_function$ sam logs -n DemoFunctionFunction --stack-name demo_function --tail
 ```
 
 You can find more information and examples about filtering Lambda function logs in the [SAM CLI Documentation](https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/serverless-sam-cli-logging.html).
@@ -100,7 +100,7 @@ You can find more information and examples about filtering Lambda function logs 
 Tests are defined in the `tests` folder in this project.
 
 ```bash
-demo-function$ ruby tests/unit/test_handler.rb
+demo_function$ ruby tests/unit/test_handler.rb
 ```
 
 ## Cleanup
@@ -108,7 +108,7 @@ demo-function$ ruby tests/unit/test_handler.rb
 To delete the sample application that you created, use the AWS CLI. Assuming you used your project name for the stack name, you can run the following:
 
 ```bash
-aws cloudformation delete-stack --stack-name demo-function
+aws cloudformation delete-stack --stack-name demo_function
 ```
 
 ## Resources
