+++ spec/identity/idp/functions_spec.rb
@@ -4,6 +4,6 @@ RSpec.describe Identity::Idp::Functions do
   end
 
   it "loads the handlers functions" do
-    expect(defined? Identity::Idp::Functions::DemoFunction::Handler).to be_truthy
+    expect(defined? Identity::Idp::Functions::DemoFunction).to be_truthy
   end
 end
